Implemented improved report format.
HTML reports now include the report data as JSON data in a JS variable. The reports now import a .js and .css file and use JQuery to present the report data.

The .js and .css file should be automatically copied to the output folder, but they can be found in the reports/ subfolder of the source distribution.